Atualizações no layout da MainActivity e na sua ViewModel; Criado o adapter para a lista de filmes; Criado as classes Repository, responsáveis pela entrega de dados à ViewModel;

+++ app/src/main/java/com/tenorinho/moviedetail/data/model/Movie.kt
@@ -8,7 +8,7 @@ data class Movie(
     @SerializedName("original_title")
     val OriginalTitle:String,
     @SerializedName("popularity")
-    val Popularity:Int,
+    val Popularity:Double,
     @SerializedName("vote_count")
     val Vote_Count:Int,
     @SerializedName("poster_path")
